refactor(catalog): the capabilities proptests move beside the model table
crates/nika-catalog/src/data/models.rs reached exactly 1500 lines once
the cached-prompt pricing landed, and the hygiene vector counts the
wall inclusively. The capabilities proptests (a `#[cfg(test)]` module
at the file's tail) move to `data/models/capabilities_proptests.rs` as
a child module — a pure move, the `super::` paths unchanged.
